requirements.txt
python-dotenv>=1.0.0
openai>=1.30.0
ccxt>=4.2.0
duckduckgo-search>=6.0.0
pandas>=2.1.0
ta>=0.11.0
streamlit>=1.35.0
